requirements.txt
flask
flask-socketio
eventlet
openai-whisper
torch
deep-translator
gtts
